Changed Octree node handling architecture
Right now, Octrees just went to a children when an AABB intersected with them. But after reading https://www.gamedev.net/articles/programming/general-and-gameplay-programming/introduction-to-octrees-r3529/
I realized that it would make more sense to have them at parent if it intersected with multiple nodes.

The issue that currently exist is that Rebuild has to be done after each insert / removal of entity.

Changes made:
- Internal nodes now will hold entities if these collide with multiple AABBs of the node children.
- Debug:
  - Added a debug utility to toggle between drawing or not drawing the Octree.
  - Colors for debugging octree are as follows:
    - If the node is a leaf and depth > 3, then color is yellow.
    - Otherwise, color will be whatever color inputted (right now green).
  - Added a Hierarchy Console Information to see how nodes and entities are distributed for ease of debugging.
- Changed the vector of the nodes to an unordered_set for efficiency and speed.
- Set the scene DEFAULT_WORLD_BOUNDS to height a bit lower than 0 so that things can be added at that height and it will not collide with all 4 original octree nodes.